Rename the project to Colorado
`andyd` is a cute play on `etcd`, but it's a little too cute.
