+++ pyproject.toml
@@ -1,11 +1,11 @@
 [tool.poetry]
-name = "andyd"
+name = "Colorado"
 version = "0.1.0"
 description = "A toy implementation of the Raft Consensus Algorithm"
 authors = ["Andy Dirnberger <andy.dirnberger@gmail.com>"]
 license = "MIT"
 readme = "README.md"
-repository = "https://github.com/dirn/andyd"
+repository = "https://github.com/dirn/Colorado"
 
 [tool.poetry.dependencies]
 python = "^3.9"
